pyproject.toml
[project]
name = "smartoj-ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "mcp>=1.15.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
